move auth routes to a separate file
