+++ server/app.js
@@ -7,33 +7,10 @@
  * https://developer.spotify.com/web-api/authorization-guide/#authorization_code_flow
  */
 const express = require('express'); // Express web server framework
-const request = require('request'); // "Request" library
 const cors = require('cors');
-const querystring = require('querystring');
 const cookieParser = require('cookie-parser');
 
-const envVars = require('./env/production');
-
-const clientId = envVars.SPOTIFY_PUBLIC; // Your client id
-const clientSecret = envVars.SPOTIFY_SECRET; // Your secret
-const redirectUri = 'http://localhost:8888/callback'; // Your redirect uri
-
-/**
- * Generates a random string containing numbers and letters
- * @param  {number} length The length of the string
- * @return {string} The generated string
- */
-const generateRandomString = (length) => {
-  const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-  let text = '';
-
-  for (let i = 0; i < length; i++) {
-    text += possible.charAt(Math.floor(Math.random() * possible.length));
-  }
-  return text;
-};
-
-const stateKey = 'spotify_auth_state';
+const spotifyAuthorizationRouter = require('./routes/spotifyAuthorization');
 
 const app = express();
 
@@ -42,114 +19,7 @@ app
   .use(cors())
   .use(cookieParser());
 
-app.get('/login', (req, res) => {
-  const state = generateRandomString(16);
-  res.cookie(stateKey, state);
-
-  // your application requests authorization
-  const scope = 'user-read-private user-read-email user-read-playback-state user-modify-playback-state';
-  res.redirect(
-    `https://accounts.spotify.com/authorize?${
-      querystring.stringify({
-        response_type: 'code',
-        client_id: clientId,
-        scope,
-        redirect_uri: redirectUri,
-        state,
-      })}`,
-  );
-});
-
-app.get('/callback', (req, res) => {
-  // your application requests refresh and access tokens
-  // after checking the state parameter
-
-  const code = req.query.code || null;
-  const state = req.query.state || null;
-  const storedState = req.cookies ? req.cookies[stateKey] : null;
-
-  if (state === null || state !== storedState) {
-    res.redirect(
-      `/#${
-        querystring.stringify({
-          error: 'state_mismatch',
-        })}`,
-    );
-  } else {
-    res.clearCookie(stateKey);
-    const authOptions = {
-      url: 'https://accounts.spotify.com/api/token',
-      form: {
-        code,
-        redirect_uri: redirectUri,
-        grant_type: 'authorization_code',
-      },
-      headers: {
-        Authorization: `Basic ${Buffer.from(`${clientId}:${clientSecret}`).toString('base64')}`,
-      },
-      json: true,
-    };
-
-    request.post(authOptions, (error, response, body) => {
-      if (!error && response.statusCode === 200) {
-        const accessToken = body.access_token;
-        const refreshToken = body.refresh_token;
-
-        const options = {
-          url: 'https://api.spotify.com/v1/me',
-          headers: { Authorization: `Bearer ${accessToken}` },
-          json: true,
-        };
-
-        // use the access token to access the Spotify Web API
-        request.get(options, (error, response, body) => {
-          console.log(body);
-        });
-
-        // we can also pass the token to the browser to make requests from there
-        res.redirect(
-          `http://localhost:3001/player/${
-            querystring.stringify({
-              access_token: accessToken,
-              refresh_token: refreshToken,
-            })}`,
-        );
-      } else {
-        res.redirect(
-          `/#${
-            querystring.stringify({
-              error: 'invalid_token',
-            })}`,
-        );
-      }
-    });
-  }
-});
-
-app.get('/refresh_token', (req, res) => {
-  // requesting access token from refresh token
-  const refreshToken = req.query.refresh_token;
-  const authOptions = {
-    url: 'https://accounts.spotify.com/api/token',
-    headers: {
-      Authorization: `Basic ${Buffer.from(`${clientId}:${clientSecret}`).toString('base64')}`,
-    },
-    form: {
-      grant_type: 'refresh_token',
-      refresh_token: refreshToken,
-    },
-    json: true,
-  };
-
-  request.post(authOptions, (error, response, body) => {
-    if (!error && response.statusCode === 200) {
-      const accessToken = body.access_token;
-      res.send({
-        access_token: accessToken,
-      });
-    }
-  });
-});
+app.use('/', spotifyAuthorizationRouter);
 
 console.log('Listening on 8888');
 app.listen(8888);
